requirements.txt
aiohttp==3.3.1
async-timeout==3.0.0
asyncinit==0.2.1
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
coverage==4.4.1
idna==2.6
idna-ssl==1.0.1
mohawk==0.3.4
multidict==4.3.1
numpy==1.14.1
pandas==0.22.0
py==1.4.34
pytest==3.2.3
pytest-cov==2.5.1
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
six==1.11.0
slugid==1.0.7
taskcluster>=4.0.0
urllib3==1.22
virtualenv==15.1.0
yarl==1.2.5
